+++ packages/react/src/hooks/useCurrentNode.ts
@@ -1,0 +1,8 @@
+import { useSelected } from './useSelected'
+import { useTree } from './useTree'
+
+export const useCurrentNode = () => {
+  const selected = useSelected()
+  const tree = useTree()
+  return tree?.findById?.(selected[0])
+}