+++ packages/react/src/hooks/useDesigner.ts
@@ -1,4 +1,4 @@
-import { useContext, useMemo } from 'react'
+import { useContext, useEffect } from 'react'
 import { Engine } from '@designable/core'
 import { DesignerContext } from '../context'
 import { isFn } from '@designable/shared'
@@ -9,9 +9,9 @@ export interface IEffects {
 export const useDesigner = (effects?: IEffects): Engine => {
   const designer: Engine =
     window['__DESINGER_ENGINE__'] || useContext(DesignerContext)?.engine
-  useMemo(() => {
+  useEffect(() => {
     if (isFn(effects)) {
-      effects(designer)
+      return effects(designer)
     }
   }, [])
   return designer